fix: 🎨 refactor and add Gif preview on node
